refactor(ui): migrate machine list tests to typescript
Migrate the test files for some machine list components to typescript.

Fixes: #2717.
