+++ ui/src/app/machines/components/ActionFormWrapper/SetZoneForm/SetZoneForm.test.tsx
@@ -1,120 +1,117 @@
-import { act } from "react-dom/test-utils";
-import { MemoryRouter, Route } from "react-router-dom";
 import { mount } from "enzyme";
+import { act } from "react-dom/test-utils";
 import { Provider } from "react-redux";
+import { MemoryRouter, Route } from "react-router-dom";
 import configureStore from "redux-mock-store";
 
-import TagForm from "./TagForm";
+import SetZoneForm from "./SetZoneForm";
+
+import type { RootState } from "app/store/root/types";
+import { NodeActions } from "app/store/types/node";
 import {
   generalState as generalStateFactory,
   machine as machineFactory,
   machineAction as machineActionFactory,
   machineActionsState as machineActionsStateFactory,
   machineState as machineStateFactory,
+  machineStatus as machineStatusFactory,
   rootState as rootStateFactory,
-  tagState as tagStateFactory,
+  zone as zoneFactory,
+  zoneState as zoneStateFactory,
 } from "testing/factories";
-import { NodeActions } from "app/store/types/node";
 
 const mockStore = configureStore();
 
-describe("TagForm", () => {
-  let initialState;
-
+describe("SetZoneForm", () => {
+  let state: RootState;
   beforeEach(() => {
-    initialState = rootStateFactory({
+    state = rootStateFactory({
       general: generalStateFactory({
         machineActions: machineActionsStateFactory({
-          data: [machineActionFactory({ name: NodeActions.TAG, title: "Tag" })],
+          data: [
+            machineActionFactory({
+              name: NodeActions.SET_ZONE,
+              title: "Set zone",
+            }),
+          ],
         }),
       }),
       machine: machineStateFactory({
+        errors: {},
+        loading: false,
         loaded: true,
         items: [
-          machineFactory({ system_id: "abc123" }),
-          machineFactory({ system_id: "def456" }),
+          machineFactory({
+            system_id: "abc123",
+          }),
+          machineFactory({
+            system_id: "def456",
+          }),
         ],
+        selected: [],
         statuses: {
-          abc123: {},
-          def456: {},
+          abc123: machineStatusFactory({}),
+          def456: machineStatusFactory({}),
         },
       }),
-      tag: tagStateFactory({
+      zone: zoneStateFactory({
+        items: [
+          zoneFactory({ id: 0, name: "default" }),
+          zoneFactory({ id: 1, name: "zone-1" }),
+        ],
         loaded: true,
       }),
     });
   });
 
-  it("dispatches action to fetch tags on load", () => {
-    const state = { ...initialState };
+  it("correctly dispatches actions to set zones of selected machines", () => {
     const store = mockStore(state);
-    mount(
-      <Provider store={store}>
-        <MemoryRouter
-          initialEntries={[{ pathname: "/machines", key: "testKey" }]}
-        >
-          <TagForm setSelectedAction={jest.fn()} />
-        </MemoryRouter>
-      </Provider>
-    );
-
-    expect(
-      store.getActions().some((action) => action.type === "tag/fetch")
-    ).toBe(true);
-  });
-
-  it("correctly dispatches actions to tag selected machines", () => {
-    const state = { ...initialState };
     state.machine.selected = ["abc123", "def456"];
-    const store = mockStore(state);
     const wrapper = mount(
       <Provider store={store}>
         <MemoryRouter
           initialEntries={[{ pathname: "/machines", key: "testKey" }]}
         >
-          <TagForm setSelectedAction={jest.fn()} />
+          <SetZoneForm actionDisabled={false} setSelectedAction={jest.fn()} />
         </MemoryRouter>
       </Provider>
     );
 
     act(() =>
-      wrapper
-        .find("Formik")
-        .props()
-        .onSubmit({
-          tags: ["tag1", "tag2"],
-        })
+      wrapper.find("Formik").props().onSubmit({
+        zone: "zone-1",
+      })
     );
     expect(
-      store.getActions().filter((action) => action.type === "machine/tag")
+      store.getActions().filter((action) => action.type === "machine/setZone")
     ).toStrictEqual([
       {
-        type: "machine/tag",
+        type: "machine/setZone",
         meta: {
           model: "machine",
           method: "action",
         },
         payload: {
           params: {
-            action: NodeActions.TAG,
+            action: NodeActions.SET_ZONE,
             extra: {
-              tags: ["tag1", "tag2"],
+              zone_id: 1,
             },
             system_id: "abc123",
           },
         },
       },
       {
-        type: "machine/tag",
+        type: "machine/setZone",
         meta: {
           model: "machine",
           method: "action",
         },
         payload: {
           params: {
-            action: NodeActions.TAG,
+            action: NodeActions.SET_ZONE,
             extra: {
-              tags: ["tag1", "tag2"],
+              zone_id: 1,
             },
             system_id: "def456",
           },
@@ -123,8 +120,7 @@ describe("TagForm", () => {
     ]);
   });
 
-  it("correctly dispatches action to tag machine from details view", () => {
-    const state = { ...initialState };
+  it("correctly dispatches action to set machine zone from details view", () => {
     state.machine.active = "abc123";
     state.machine.selected = [];
     const store = mockStore(state);
@@ -136,35 +132,36 @@ describe("TagForm", () => {
           <Route
             exact
             path="/machine/:id"
-            component={() => <TagForm setSelectedAction={jest.fn()} />}
+            component={() => (
+              <SetZoneForm
+                actionDisabled={false}
+                setSelectedAction={jest.fn()}
+              />
+            )}
           />
         </MemoryRouter>
       </Provider>
     );
 
     act(() =>
-      wrapper
-        .find("Formik")
-        .props()
-        .onSubmit({
-          tags: ["tag1", "tag2"],
-        })
+      wrapper.find("Formik").props().onSubmit({
+        zone: "zone-1",
+      })
     );
-
     expect(
-      store.getActions().filter((action) => action.type === "machine/tag")
+      store.getActions().filter((action) => action.type === "machine/setZone")
     ).toStrictEqual([
       {
-        type: "machine/tag",
+        type: "machine/setZone",
         meta: {
           model: "machine",
           method: "action",
         },
         payload: {
           params: {
-            action: NodeActions.TAG,
+            action: NodeActions.SET_ZONE,
             extra: {
-              tags: ["tag1", "tag2"],
+              zone_id: 1,
             },
             system_id: "abc123",
           },
